+++ src/modules/mavlink/mavlink_messages.cpp
@@ -112,6 +112,7 @@ using matrix::wrap_2pi;
 #include "streams/DISTANCE_SENSOR.hpp"
 #include "streams/ESC_INFO.hpp"
 #include "streams/ESC_STATUS.hpp"
+#include "streams/ESTIMATOR_STATUS.hpp"
 #include "streams/EXTENDED_SYS_STATE.hpp"
 #include "streams/FLIGHT_INFORMATION.hpp"
 #include "streams/GPS_GLOBAL_ORIGIN.hpp"
@@ -2787,87 +2788,6 @@ protected:
 	}
 };
 
-class MavlinkStreamEstimatorStatus : public MavlinkStream
-{
-public:
-	const char *get_name() const override
-	{
-		return MavlinkStreamEstimatorStatus::get_name_static();
-	}
-
-	static constexpr const char *get_name_static()
-	{
-		return "ESTIMATOR_STATUS";
-	}
-
-	static constexpr uint16_t get_id_static()
-	{
-		return MAVLINK_MSG_ID_ESTIMATOR_STATUS;
-	}
-
-	uint16_t get_id() override
-	{
-		return get_id_static();
-	}
-
-	static MavlinkStream *new_instance(Mavlink *mavlink)
-	{
-		return new MavlinkStreamEstimatorStatus(mavlink);
-	}
-
-	unsigned get_size() override
-	{
-		return _estimator_status_sub.advertised() ? MAVLINK_MSG_ID_ESTIMATOR_STATUS_LEN + MAVLINK_NUM_NON_PAYLOAD_BYTES : 0;
-	}
-
-private:
-	uORB::Subscription _estimator_selector_status_sub{ORB_ID(estimator_selector_status)};
-	uORB::Subscription _estimator_status_sub{ORB_ID(estimator_status)};
-
-	/* do not allow top copying this class */
-	MavlinkStreamEstimatorStatus(MavlinkStreamEstimatorStatus &) = delete;
-	MavlinkStreamEstimatorStatus &operator = (const MavlinkStreamEstimatorStatus &) = delete;
-
-protected:
-	explicit MavlinkStreamEstimatorStatus(Mavlink *mavlink) : MavlinkStream(mavlink)
-	{}
-
-	bool send() override
-	{
-		// use primary estimator_status
-		if (_estimator_selector_status_sub.updated()) {
-			estimator_selector_status_s estimator_selector_status;
-
-			if (_estimator_selector_status_sub.copy(&estimator_selector_status)) {
-				if (estimator_selector_status.primary_instance != _estimator_status_sub.get_instance()) {
-					_estimator_status_sub.ChangeInstance(estimator_selector_status.primary_instance);
-				}
-			}
-		}
-
-		estimator_status_s est;
-
-		if (_estimator_status_sub.update(&est)) {
-			mavlink_estimator_status_t est_msg{};
-			est_msg.time_usec = est.timestamp;
-			est_msg.vel_ratio = est.vel_test_ratio;
-			est_msg.pos_horiz_ratio = est.pos_test_ratio;
-			est_msg.pos_vert_ratio = est.hgt_test_ratio;
-			est_msg.mag_ratio = est.mag_test_ratio;
-			est_msg.hagl_ratio = est.hagl_test_ratio;
-			est_msg.tas_ratio = est.tas_test_ratio;
-			est_msg.pos_horiz_accuracy = est.pos_horiz_accuracy;
-			est_msg.pos_vert_accuracy = est.pos_vert_accuracy;
-			est_msg.flags = est.solution_status_flags;
-			mavlink_msg_estimator_status_send_struct(_mavlink->get_channel(), &est_msg);
-
-			return true;
-		}
-
-		return false;
-	}
-};
-
 class MavlinkStreamCameraCapture : public MavlinkStream
 {
 public:
@@ -2977,7 +2897,9 @@ static const StreamListItem streams_list[] = {
 	create_stream_list_item<MavlinkStreamGlobalPositionInt>(),
 	create_stream_list_item<MavlinkStreamLocalPositionNED>(),
 	create_stream_list_item<MavlinkStreamOdometry>(),
+#if defined(ESTIMATOR_STATUS_HPP)
 	create_stream_list_item<MavlinkStreamEstimatorStatus>(),
+#endif // ESTIMATOR_STATUS_HPP
 #if defined(VIBRATION_HPP)
 	create_stream_list_item<MavlinkStreamVibration>(),
 #endif // VIBRATION_HPP
